feat: add user mention system

+++ Tg_Bot/Migrations/20250818153142_username_db.cs
@@ -16,7 +16,7 @@ namespace Tg_Bot.Migrations
                 name: "Users",
                 columns: table => new
                 {
-                    UserId = table.Column<int>(type: "integer", nullable: false)
+                    UserId = table.Column<long>(type: "long", nullable: false)
                         .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
                     UserName = table.Column<string>(type: "text", nullable: false),
                     RealName = table.Column<string>(type: "text", nullable: false),
