+++ Tg_Bot/Migrations/ApplicationContextModelSnapshot.cs
@@ -25,11 +25,11 @@ namespace Tg_Bot.Migrations
 
             modelBuilder.Entity("Tg_Bot.Models.Users", b =>
                 {
-                    b.Property<int>("UserId")
+                    b.Property<long>("UserId")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("long");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("UserId"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("UserId"));
 
                     b.Property<string>("RealName")
                         .IsRequired()
